Now compiles and some functions work
I've monkey'd around with the memory and it is sployy! I'm using 7300
for the $A000 memory, 7400 for the $D000 and $4xxx memory. I still
have a lot of fixing up of the memory. I've managed to get it to start
under sim6800. HE, !!, RE(?) and ME work. Main commands loop now works. PU
and LO are missing. I'll probably steal code from Mikbug.asm for now.
It's still a mess but it's getting there.